Refactor `execution.rs` to fix SRP violation by splitting out context builder, batch runner, and session access.
